+++ BetterFlashcards/App/Services/Protocols/FlashCardServiceProtocol.swift
@@ -11,13 +11,11 @@ import Combine
 protocol FlashCardServiceProtocol {
     var eventPublisher: AnyPublisher<BasicServiceEvent<FlashCard>, Never> { get }
 
-    func getList(for deckID: Deck.ID) async throws -> [FlashCard]
     func getList(for deckID: Deck.ID, at: Pagination) async throws -> PaginatedList<FlashCard>
     func create(_: FlashCard) async throws -> FlashCard
     func update(_: FlashCard) async throws -> FlashCard
     func delete(_: FlashCard) async throws -> FlashCard
     
-    func dueList(for deckID: Deck.ID) async throws -> [FlashCard]
     func dueList(for deckID: Deck.ID, at: Pagination) async throws -> PaginatedList<FlashCard>
     func save(practice: FlashCardPractice) async throws
 }
