+++ BetterFlashcards/Infrastructure/Repositories/Protocols/DeckRepositoryProtocol.swift
@@ -8,7 +8,6 @@
 import Foundation
 
 protocol DeckRepositoryProtocol {
-    func fetchAll() async throws -> [Deck]
     func fetch(at: Pagination) async throws -> PaginatedList<Deck>
     func fetch(by deckID: Deck.ID) async throws -> Deck?
     func create(deck: Deck) async throws -> Deck
