+++ BetterFlashcards/Infrastructure/Repositories/Protocols/FlashCardRepositoryProtocol.swift
@@ -8,14 +8,12 @@
 import Foundation
 
 protocol FlashCardRepositoryProtocol {
-    func fetchAll(by deckID: Deck.ID) async throws -> [FlashCard]
     func fetch(by deckID: Deck.ID, at: Pagination) async throws -> PaginatedList<FlashCard>
     func fetch(by flashCardID: FlashCard.ID) async throws -> FlashCard?
     func create(flashCard: FlashCard) async throws -> FlashCard
     func update(flashCard: FlashCard) async throws -> FlashCard
     func delete(flashCard: FlashCard) async throws
-    
-    func dueCards(for deckID: Deck.ID) async throws -> [FlashCard]
+
     func dueCards(for deckID: Deck.ID, at: Pagination) async throws -> PaginatedList<FlashCard>
     func add(practice: FlashCardPractice) async throws
 }
